requirements.txt
google-generativeai
keyboard
pyperclip
colorama
python-dotenv
